// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  TIPSTER: 'TIPSTER',
  BETTOR: 'BETTOR'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PickStatus = {
  PENDING: 'PENDING',
  LOCKED: 'LOCKED',
  WON: 'WON',
  LOST: 'LOST',
  VOID: 'VOID'
} as const

export type PickStatus = (typeof PickStatus)[keyof typeof PickStatus]


export const SubStatus = {
  ACTIVE: 'ACTIVE',
  CANCELLED: 'CANCELLED',
  EXPIRED: 'EXPIRED'
} as const

export type SubStatus = (typeof SubStatus)[keyof typeof SubStatus]


export const PayStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED'
} as const

export type PayStatus = (typeof PayStatus)[keyof typeof PayStatus]